Move ty lowering diagnostic definitions into a separate module
To keep them organized.
